+++ test/must/own_properties_test.js
@@ -1,0 +1,3 @@
+describe("Must.prototype.ownProperties", function() {
+  require("./_properties_test")("ownProperties", false)
+})